+++ lib/neopixel/Neopixel_Blocking.cpp
@@ -1,4 +1,4 @@
-#include "BlockingNeopixel.h"
+#include "Neopixel_Blocking.h"
 #include "mbed.h"
 
 // constexpr T0H
